LayoutTests/fast/events/ios/did-not-handle-meaningful-click-expected.txt

This test is intended to exercise the 'meaningful click' heuristic, and requires WebKitTestRunner.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS didHandleTap is true
PASS didHandleTapAsMeaningfulClick is false
PASS clicked on button
PASS didHandleTap is true
PASS didHandleTapAsMeaningfulClick is true
PASS didHandleTap is true
PASS didHandleTapAsMeaningfulClick is false
PASS didHandleTap is true
PASS didHandleTapAsMeaningfulClick is true
PASS successfullyParsed is true

TEST COMPLETE
